refactor: move getPage to an util file
